+++ src/app/ngx-editor/ngx-editor.module.ts
@@ -4,7 +4,7 @@ import { NgxEditorComponent } from './ngx-editor.component';
 import { NgxGrippieComponent } from './ngx-grippie/ngx-grippie.component';
 import { NgxToolbarComponent } from './ngx-toolbar/ngx-toolbar.component';
 import { NgxEditorMessageComponent } from './ngx-editor-message/ngx-editor-message.component';
-import { MessageService } from './ngx-editor-message/message.service';
+import { MessageService } from './common/services/message.service';
 import { CommandExecutor } from './common/services/command-executor';
 
 @NgModule({
